Fix KPluginSelector not adding non .desktop file plugins
KPluginSelector always assumed those plugins had already been added to
the list because it only compares KPluginInfo::entryPath() for equality
which returns an empty string if the plugin metadata was loaded directly
from the shared library. Therefore as soon as one plugin with the
metadata embedded into the library had been loaded all others would
compare equal to that.

REVIEW: 120586

Is this the correct fix or should KPluginInfo::entryPath() be changed?