[KCModuleQml] Ignore disabled controls when tabbing
Whether an item is enabled or not doesn't seem to affect its focusPolicy or position in the focus chain.
Explicitly ignore disabled controls and move to the next.
Fixes being unable to tab out of a QML KCM when the "Help" button (typically next control after the view in System Settings) is disabled

Differential Revision: https://phabricator.kde.org/D12955