Supress mouse events in KCMs causing window moves
Summary:
If a mouse press is handled by the new input handlers the mouse evnet is
not accepted at a window level.

This causes the breeze style to start a window drag.

Marking all mouse events as accepted after being processed supresses
them reaching the widget.

This is noticable in the rewritten kscreen KCM when running on XCB. (it
would also affect wayland, but the breeze window management is currently
broken there)

Test Plan: Dragged a monitor in kscreen kcm. The monitor representation moved not the whole window

Reviewers: #plasma

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D24012