Fix memory leak of KQuickAddons::ConfigModule objects
Summary:
When we create a KQuickAddons::ConfigModule  in the factory we didn't
set a parent. It is passed to the wrapper KCModuleQML, and even though
the wrapper is memory manager the ConfigModule object itself was not.

This lead to another crash as the KAboutData is deleted twice, which was
fixed with an explicit copy

BUG: 412998

Test Plan:
qDebug inside the colours KCM destructor
It now appears when changing tabs

Reviewers: #plasma

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D24716