Use KCModuleInfo::service() to load modules, for both KPluginInfo and KService code paths
This is an addendum to commit 53b41bc90b354ce7642; since now we have two
code paths, KPluginInfo and KService, instead of using pluginInfo(), use
module.service(), because the latter is going to return the KService associated
with the module, from either code path.

This lets systemsettings5 work with loading KCModules that have an
actual library, and the special case where systemsettings creates a fake
kcm that just loads an external app.