Move the KCMUtilsGenerateModuleData to correct place
Also rename this file, let's follow the naming conventions used by
frameworks.

File should not be installed in ECM path, but rather in kcmutils
specific path, also include this file from the KCMUtilsConfig.cmake
file, which makes it automatically include it if find_package call is
placed.

This is more in line with how macros are handled