kcmoduleqml: don't bind column width to view width
This causes a lot of heisenbugs due to bindings triggering at inconsistent times, sometimes causing a KCM's requested width to be respected, and other times, causing it to be ignored. Since we're going to be setting this from C++ code, we don't need to create a QML binding for this.

BUG: 428727