KPluginSelector: Allow KCMs to be loaded using X-KDE-ConfigModule keyword
This way providers of the KCM don't need to install a desktop file for the KCM
and only needs to define the key in order for the KCM to be picked up.

This way the conversion of the desktop files is no longer blocked (https://phabricator.kde.org/T14564)
and providers don't have to do any changes when KPluginSelector is ported away from KPluginInfo.
See https://phabricator.kde.org/T12265.

The key is already specified in https://phabricator.kde.org/T13555#248671.