Implement KPluginSelector replacement not based around KPluginInfo
This is a KPluginMetaData based replacement for KPluginSelector, which
internally uses the KPluginModel to allow the creation of a QML variant.

This way consumers can fully port away from KPluginInfo in their code.
By using KServiceTypeTrader for the deprecated code path we can keep existing
KCMs of plugins working. In case of KRunner this would even be third party KCMs.

This work is based on the initial draft of Nicolas Fella.

Task: https://phabricator.kde.org/T12265