KCModuleProxy: Use std::optional to check if we have a KPluginMetaData object
For KCMs that are loaded as part of a KPluginSelector/KPluginWidget we don't need json metadata.

For plugins that are dynamically loaded we insert a dummy json object, but in case of static plugins
we do not have an overload to do so. Meaning consumers are forced to provide json metadata for the plugin
to be considered valid. See https://phabricator.kde.org/T15094 which will provide a more universal way of handling it.