kpluginmodel: Also copy plugin Id from KService to json object
In almost all cases we correctly derive the id from the filename
or it does not matter. However in KWin we have a central factory plugin
which loads the xml/ui files for the other plugins. Here the plugin id is needed
to determine the location for the xml/ui files.

Without copying the json key, the plugin id would point to the factory plugin.

The idea behind not using the json conversion code was to not introduce
new usage of deprecated API.