KCModuleLoader: repair loading KCModules
This fixes `kcmshell5 fontinst` for me. Because it's not a
KQuickAddons::ConfigModule, the call to KPluginFactory::instantiatePlugin
fails, but we still want to proceed and try loading it as a KCModule.

Not sure why this requires going into the KService-related code path
though, surely we can load a KCModule via JSON?