KPluginModel: Add method to get KPluginMetaData config object for given plugin id
Unlike KPluginSelector/KPluginWidget, the KPluginModel does not have the functionality to
display the config of a plugin. Meaning consumers will need to take care of implementing a `showConfiguration` like method.
For doing this we need the KPluginMetaData object of the config.

For example the KRunner KCM has the functionality to open a plugin's config by passing in the
plugin id to the args when creating the KCModule.