KPluginSelector: Remove unnecessary background rect in About sheet
It worsens text contrast and causes visual bugs because it's being
set as the background of the entire Sheet rather than its content
item, which makes it hide the sheet's nice rounded corners. We could
fix that by setting it on the content item, removing internal padding
from the Sheet itself, and then manually re-adding it in the content
item, but that would be a lot of code to preserve something that has
less than ideal text contrast anyway, so let's just kill two birds
with one stone by removing the background color rectangle entirely.

BUG: 457648
FIXED-IN: 5.98