KCModuleData: Remove virtual hook
This would be useful as a placeholder for future extensions that otherwise break BIC.
However, this class is only a wrapper and does not contain much logic.
In case we need to add a new method that consumers can overwrite, we could use
QMetaObject to call this method. This is cleaner on the consumer side too.