Fix KCMultiDialog not saving modified configs
clientChanged() gets called when needsSave() changes, for example when
the user has changed an option.

On the other hand, clientChanged() breaks the connection between the
apply button and its associated slot. In other words, the OK and the
Apply button will become dysfunctional the moment user changes an
option. It looks like some sort of oversight.