KCModuleQml: Load KCMs QML files using qrc
This way we do not need KPackage anymore and the plugins contain all files that they need.
Meaning one can not test KCMs without having to install them.

Also, we ensure at build-time that all the intended files are shipped with the plugin.
Other KPackage features like the changeable mainscript file are no longer supported, but
renaming the entry-point qml file to sth. else that is specified in the metadata is not ideal anyway.
Since most QML KCMs do not ship metadata for the package (see https://phabricator.kde.org/T14983),
this should affect very few usages.

KPackage was originally used, because it was considered to have QML-only KCMs.
However, this practically never happened and there was no need to do so.