Provide dummy incubation controller
QQmlIncubator can act synchronously when it's asked to create a
component asynchronously if the QML engine has missing incubation
controller.

It's typically not a problem because we get a QML engine from
QQuickWindow or QQuickView, whose incubation controller is very smart.

On the other hand, Kirigami code relies on the fact that asynchronous
QQmlIncubator will always act asynchronously. Otherwise it acts quite
bad to the point where it hits an infinite recursion.

In order to make QtQuick based config modules work, this change adds a
dummy incubation controller. It's not great, and we need to revisit
whether having a shared qml engine is the right approach towards
providing the support for QML based KCMs.