[kpluginwidget] Fix parent handling for plugin KCMs
We add the KCM to a layout, so we should not set a parent ourselves before
adding the widget to the layout.

Setting a parent to the KCM directly in fact breaks the layout and size handling.

In KF5 this didn't happen because there was a KCModuleProxy between the actual KCM and
the layout. Add a dummy QWidget to emulate that and fix the issue