kcmshell: Remove dbus registration of modules
This is quite problematic, because in case modules are already launched,
it will print a message on the console and wait for the other instance
to finish.
Also, the registration was fragile, because in case multiple KCMs are
loaded, the servicename would be different. Meaning if you launch the
KCMs "icons" and "colors" in one instance and then "colors" and "icons"
in another one, you would have to kcmshell instances displaying the same
KCMs.

Because we use systemsettings (which is a single-instance app) for most KCMs,
it is pretty unlikely that two KCMs are opened at the same time anyway.