KAbstractConfigModule: Allow updating arguments on a running module
Previously, arguments were only passed to the module in the
constructor, which limited subsequent calls on an module which
was already running.

This adds a new `activationRequested()` signal with the `args`
parameter to make updating arguments possible. The KCM will
connect to this signal if it needs to process then, for instance
to move to a specific page or item.