GridDelegate: Fix ring color for disabled items
Stop showing a coloured ring for the hovered items when
they are disabled.

The current selected item still shows a ring becase, but just
with the regular highlight color, because the neutral highlight
doesn't get get desaturated when hovered.

While at it, simplify the cascading conditions logic and use
the newer `Qt.alpha()` function.