Introduce the sidebar mode for kcms
There are some kcms (gamepad, users) that have a page that acts like a
sidebar, which should go fullscreen and be its own sliding page in plasma
mobile. There seems to be the same code copied and pasted in several
kcms. this changes making a single property sidebarMode in AbstractKCM
making client code much cleaner