Add API to make header and footer paddings optional
This facilitates KCMs that want to touch the edges of the page with
header or footer content. For example, an InlineMessage with the
position property set to "header."

SimpleKCM only gets the header property since it doesn't have special
handling for a footer. Also for SimpleKCM the padding defaults to off
rather than on, replicating the status quo (which appears to be the
result of a change we forgot to make for KF6).