GridDelegate: Don't use qualified property access as appropriate
At the top-level component there should be no need to use qualified
property access for the properties of the same component. In fact, this
should be true for any component on any level, not only top-level ones.
Especially if the `return` lines in this block already use unqualified
access for the very same properties.