QuickModuleLoader: Use QML engine from application if no engine was provided
We want to try and prevent creating a new engine if at all possible. For
System Settings, we now set a "__qmlEngine" property on the application.
Should KQuickConfigModuleLoader::loadModule() be called without an
engine paramter, first try and see if there is an application-wide
engine set on the application and use that if so, rather than only using
an internal shared engine.