Always show title and caption in tooltip if truncated
We only show title and caption in tooltip if they are truncated and
no explicit tooltip is set. However, this means the truncated title
and caption are never showed in full whenever tooltip is set.
Always join the truncated title and caption in the tooltip.

I discovered this when configuring virtual keyboard. The title for
fcitx is long, and not able to see the full title is very
frustrating.