GridDelegate: cap label line counts to 1
This was the intention all along, but thus far was accomplished
implicitly rather than explicitly, by limiting the total height of the
parent layout. That usually works, but can break when either of the
labels are fed a specially-crafted very long string. In such a case, the
layout breaks horribly.

This commit fixes that issue by setting the maximum line count values,
making the implicit explicit.


BUG: 497253
FIXED-IN: 6.10