ScrollView: don't move focus to empty toolbar
ScrollViewKCM uses a custom ScrollView component that binds a keyboard
handler to the up arrow key, passing focus to the page's toolbar if the
ScollView's content does not handle it. This is convenient to reach the
actions in the toolbar (such as the "Add New…" action in kcm_users), but
it is pointless in case the toolbar is empty - focus is passed to a
toolbar that has no contents and doesn't do anything, so to the user it
appears as if keyboard focus has vanished into thin air.

This change make the KeyNavigation target for the up arrow key
conditional on there being any actions set for the toolbar.