kcmoduleqml: simplify and improve focus handling
kcmoduleqml contains some old tricks to make passing focus between the
qml and qtwidgets bits, but they are fragile, and break in many cases -
especially with systemsettings adding another layer of qml. The hacks
may have been necessary in the past, but things actually work well now,
so we can essentially just set a focusProxy and be done ... as long as
we don't care about accessibility. The qml bits receive focus before the
widgets parts fully hand it over, so the first qml item receiving focus
will not be  announced over screen readers. We do care though, and thus
some special handling is still needed.

We switch to using focusProxy, and add a bit of surgery to the focus
transition so that screen readers always stay in the loop about what
things are focused. We also need a bit of special handling for the
backtab case, which doesn't quite work automatically with
activeFocusOnTab set on the qml root.