kcmultidialog: set AccessibleName for ScrollArea
The tab stop at the kcm scrollarea can in some cases serve a purpose,
but is very cryptic to screen reader users as it appears to be non-
interactable and is only announced with the confusing "LayeredPane".

Instead set an explicit AccessibleName that informs the user of the
general purpose of the control they are on. QAccessible does not support
the Atspi role SCROLL_PANE which would be optimal here, and there is no
clear equivalent, so it has to be done with AccessibleName.