GridDelegate & GridViewInternal: Drop mobile tweaks
This commit drops mobile tweaks the sizing algorithm for gridviews, so
that the delegates have unified behavior between both mobile and desktop.

This also globally lowers the minimum number of gridview columns to 1
(even for desktop, which used to be 2) so that it is the same for both
mobile and desktop. I think this is fine for desktop too, so that
delegates don't get crammed if the view is very narrow.